+++ src/pixels/comet.h
@@ -1,3 +1,3 @@
 #include <Adafruit_NeoPixel.h>
 
-void updateComet(Adafruit_NeoPixel& pixels, uint8_t state[]);
+void updateComet(Adafruit_NeoPixel& pixels, uint16_t state[]);
