+++ src/pixels/firework.h
@@ -1,3 +1,3 @@
 #include <Adafruit_NeoPixel.h>
 
-void updateFirework(Adafruit_NeoPixel& pixels, uint8_t state[]);
+void updateFirework(Adafruit_NeoPixel& pixels, uint16_t state[]);
