+++ src/pixels/rain.h
@@ -1,3 +1,3 @@
 #include <Adafruit_NeoPixel.h>
 
-void updateRain(Adafruit_NeoPixel& pixels, uint8_t state[]);
+void updateRain(Adafruit_NeoPixel& pixels, uint16_t state[]);
