+++ src/pixels/spiral.h
@@ -1,3 +1,3 @@
 #include <Adafruit_NeoPixel.h>
 
-void updateSpiral(Adafruit_NeoPixel& pixels, uint8_t state[]);
+void updateSpiral(Adafruit_NeoPixel& pixels, uint16_t state[]);
